backend/pyproject.toml
[project]
name = "live-sales-coach"
version = "0.1.0"
description = "Real-time AI Sales Coach powered by Gemini Live API + Google ADK"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.25.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.28.0",
    "google-cloud-firestore>=2.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
